pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "MadNIS-NLO"
version = "1.0.0"
description = "MadNIS at NLO"
authors = [
    { name="Giovanni De Crescenzo", email="crescenzo@thphys.uni-heidelberg.de" },
    { name="Javier Marino Villadamigo", email="marino@thphys.uni-heidelberg.de" },
    { name="Theo Heimel", email="theo.heimel@uclouvain.be" },
    { name="Ramon Winterhalder", email="ramon.winterhalder@unimi.it" },
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
urls = {"Repository" = "https://github.com/heidelberg-hepml/parx"}
dependencies = [
    "torch>=2.2",
    "numpy<2.0",
    "matplotlib",
    "pyyaml",
    "scipy",
    "black==22.3.0",
    "vegas",
    "fastjet",
    "madspace",
    "torchspace @ git+https://github.com/madgraph-ml/torchspace.git",
    "madnis",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "tqdm",
    "tensorboard==2.20.0",
]

[tool.black]
line-length = 90

[tool.isort]
profile = "black"
atomic = true
line_length = 90